+++ src/eos/eosEngine.js
@@ -5,6 +5,7 @@ import { bns } from 'biggystring'
 import {
   type EdgeCurrencyEngineOptions,
   type EdgeCurrencyTools,
+  type EdgeFetchFunction,
   type EdgeFreshAddress,
   type EdgeSpendInfo,
   type EdgeTransaction,
@@ -12,7 +13,10 @@ import {
   InsufficientFundsError,
   NoAmountSpecifiedError
 } from 'edge-core-js/types'
-import eosjs from 'eosjs'
+import { Api, JsonRpc, RpcError } from 'eosjs'
+import EosApi from 'eosjs-api'
+import { JsSignatureProvider } from 'eosjs/dist/eosjs-jssig'
+import { convertLegacyPublicKeys } from 'eosjs/dist/eosjs-numeric'
 
 import { CurrencyEngine } from '../common/engine.js'
 import {
@@ -43,8 +47,38 @@ type EosFunction =
   | 'getInfo'
   | 'getKeyAccounts'
   | 'getOutgoingTransactions'
-  | 'transaction'
+  | 'transact'
 
+class CosignAuthorityProvider {
+  rpc: JsonRpc
+  constructor(rpc: JsonRpc) {
+    this.rpc = rpc
+  }
+
+  async getRequiredKeys(args) {
+    const { transaction } = args
+    // Iterate over the actions and authorizations
+    transaction.actions.forEach((action, ti) => {
+      action.authorization.forEach((auth, ai) => {
+        // If the authorization matches the expected cosigner
+        // then remove it from the transaction while checking
+        // for what public keys are required
+        if (auth.actor === 'greymassfuel' && auth.permission === 'cosign') {
+          delete transaction.actions[ti].authorization.splice(ai, 1)
+        }
+      })
+    })
+    // the rpc below should be an already configured JsonRPC client from eosjs
+    return convertLegacyPublicKeys(
+      (
+        await this.rpc.fetch('/v1/chain/get_required_keys', {
+          transaction,
+          available_keys: args.availableKeys
+        })
+      ).required_keys
+    )
+  }
+}
 export class EosEngine extends CurrencyEngine {
   // TODO: Add currency specific params
   // Store any per wallet specific data in the `currencyEngine` object. Add any params
@@ -59,7 +93,7 @@ export class EosEngine extends CurrencyEngine {
     currencyPlugin: EosPlugin,
     walletInfo: EdgeWalletInfo,
     opts: EdgeCurrencyEngineOptions,
-    fetchJson: Function
+    fetchCors: EdgeFetchFunction
   ) {
     super(currencyPlugin, walletInfo, opts)
 
@@ -100,8 +134,12 @@ export class EosEngine extends CurrencyEngine {
         }
         const eosPaymentServer = this.currencyInfo.defaultSettings.otherSettings
           .eosActivationServers[0]
-        const url = `${eosPaymentServer}/api/v1/activateAccount`
-        return fetchJson(url, options)
+        const uri = `${eosPaymentServer}/api/v1/activateAccount`
+        const response = await fetchCors(uri, options)
+        if (!response.ok) {
+          throw new Error(`Error ${response.status} while fetching ${uri}`)
+        }
+        return response.json()
       }
     }
   }
@@ -201,13 +239,13 @@ export class EosEngine extends CurrencyEngine {
     const date = Date.parse(action['@timestamp']) / 1000
     const blockHeight = action.block_num > 0 ? action.block_num : 0
     if (!action.block_num) {
-      this.log('Invalid EOS transaction data. No tx block_num')
+      this.log('Invalid transaction data. No tx block_num')
       return 0
     }
     const txid = action.trx_id
 
     if (!action.act) {
-      this.log('Invalid EOS transaction data. No action.act')
+      this.log('Invalid transaction data. No action.act')
       return 0
     }
     const name = action.act.name
@@ -216,7 +254,7 @@ export class EosEngine extends CurrencyEngine {
     // this.log(`Action type: ${name}`)
     if (name === 'transfer') {
       if (!action.act.data) {
-        this.log('Invalid EOS transaction data. No action.act.data')
+        this.log('Invalid transaction data. No action.act.data')
         return 0
       }
       const { from, to, memo, amount, symbol } = action.act.data
@@ -442,14 +480,47 @@ export class EosEngine extends CurrencyEngine {
       }
 
       case 'getCurrencyBalance':
-      case 'getInfo':
-      case 'transaction': {
+      case 'getInfo': {
         const { eosNodes } = this.currencyInfo.defaultSettings.otherSettings
         const randomNodes = pickRandom(eosNodes, 3)
         out = await asyncWaterfall(
           randomNodes.map(server => async () => {
-            const eosServer = eosjs({ ...eosConfig, httpEndpoint: server })
+            const eosServer = EosApi({ ...eosConfig, httpEndpoint: server })
             const result = await eosServer[func](...params)
+            return { server, result }
+          })
+        )
+        break
+      }
+      case 'transact': {
+        const {
+          eosFuelServers
+        } = this.currencyInfo.defaultSettings.otherSettings
+        const randomNodes = pickRandom(eosFuelServers, 30)
+        out = await asyncWaterfall(
+          randomNodes.map(server => async () => {
+            const rpc = new JsonRpc(server, {
+              fetch: (...args) => {
+                // this.log(`LoggedFetch: ${JSON.stringify(args)}`)
+                return eosConfig.fetch(...args)
+              }
+            })
+            const keys = params[1].keyProvider ? params[1].keyProvider : []
+            params[1] = {
+              ...params[1],
+              blocksBehind: 3,
+              expireSeconds: 30
+            }
+            const signatureProvider = new JsSignatureProvider(keys)
+            const eos = new Api({
+              // Pass in new authorityProvider
+              authorityProvider: new CosignAuthorityProvider(rpc),
+              rpc,
+              signatureProvider,
+              textDecoder: new TextDecoder(),
+              textEncoder: new TextEncoder()
+            })
+            const result = await eos[func](...params)
             return { server, result }
           })
         )
@@ -457,7 +528,7 @@ export class EosEngine extends CurrencyEngine {
       }
     }
 
-    this.log(`EOS multicastServers ${func} ${out.server} won`)
+    this.log(`multicastServers ${func} ${out.server} won`)
     return out.result
   }
 
@@ -628,8 +699,8 @@ export class EosEngine extends CurrencyEngine {
     if (bns.gt(nativeAmount, nativeBalance)) {
       throw new InsufficientFundsError()
     }
-    const DecimalPad = eosjs.modules.format.DecimalPad
-    const quantity = DecimalPad(exchangeAmount, 4) + ` ${currencyCode}`
+
+    const quantity = bns.toFixed(exchangeAmount, 4, 4) + ` ${currencyCode}`
     let memo = ''
     if (
       edgeSpendInfo.spendTargets[0].otherParams &&
@@ -638,8 +709,20 @@ export class EosEngine extends CurrencyEngine {
     ) {
       memo = edgeSpendInfo.spendTargets[0].otherParams.uniqueIdentifier
     }
+
     const transactionJson = {
       actions: [
+        {
+          authorization: [
+            {
+              actor: 'greymassfuel',
+              permission: 'cosign'
+            }
+          ],
+          account: 'greymassnoop',
+          name: 'noop',
+          data: {}
+        },
         {
           account: 'eosio.token',
           name: 'transfer',
@@ -659,11 +742,12 @@ export class EosEngine extends CurrencyEngine {
       ]
     }
 
+    // XXX Greymass doesn't let us hit their servers too often
     // Create an unsigned transaction to catch any errors
-    await this.multicastServers('transaction', transactionJson, {
-      sign: false,
-      broadcast: false
-    })
+    // await this.multicastServers('transact', transactionJson, {
+    //   sign: false,
+    //   broadcast: false
+    // })
 
     nativeAmount = `-${nativeAmount}`
 
@@ -682,7 +766,7 @@ export class EosEngine extends CurrencyEngine {
       }
     }
 
-    this.log('EOS transaction prepared')
+    this.log('transaction prepared')
     this.log(
       `${nativeAmount} ${this.walletLocalData.publicKey} -> ${publicAddress}`
     )
@@ -745,23 +829,24 @@ export class EosEngine extends CurrencyEngine {
   // }
 
   async signTx(edgeTransaction: EdgeTransaction): Promise<EdgeTransaction> {
-    const otherParams = getOtherParams(edgeTransaction)
+    // const otherParams = getOtherParams(edgeTransaction)
 
     // Do signing
     // Take the private key from this.walletInfo.keys.eosKey and sign the transaction
     // const privateKey = this.walletInfo.keys.eosKey
-    const keyProvider = []
-    if (this.walletInfo.keys.eosKey) {
-      keyProvider.push(this.walletInfo.keys.eosKey)
-    }
-    if (this.walletInfo.keys.eosOwnerKey) {
-      keyProvider.push(this.walletInfo.keys.eosOwnerKey)
-    }
-    await this.multicastServers('transaction', otherParams.transactionJson, {
-      keyProvider,
-      sign: true,
-      broadcast: false
-    })
+    // const keyProvider = []
+    // if (this.walletInfo.keys.eosKey) {
+    //   keyProvider.push(this.walletInfo.keys.eosKey)
+    // }
+    // if (this.walletInfo.keys.eosOwnerKey) {
+    //   keyProvider.push(this.walletInfo.keys.eosOwnerKey)
+    // }
+    // XXX Greymass doesn't let us hit their servers too often
+    // await this.multicastServers('transact', otherParams.transactionJson, {
+    //   keyProvider,
+    //   sign: true,
+    //   broadcast: false
+    // })
 
     // Complete edgeTransaction.txid params if possible at this state
     return edgeTransaction
@@ -782,7 +867,7 @@ export class EosEngine extends CurrencyEngine {
     }
     try {
       const signedTx = await this.multicastServers(
-        'transaction',
+        'transact',
         otherParams.transactionJson,
         {
           keyProvider,
@@ -794,7 +879,13 @@ export class EosEngine extends CurrencyEngine {
       edgeTransaction.txid = signedTx.transaction_id
       return edgeTransaction
     } catch (e) {
+      this.log('\nCaught exception: ' + e)
+      if (e instanceof RpcError) this.log(JSON.stringify(e.json, null, 2))
       let err = e
+      if (err.error) {
+        this.log(`err.error= ${err.error}`)
+        this.log(`err.error.name= ${err.error.name}`)
+      }
       try {
         err = JSON.parse(e)
       } catch (e2) {