+++ src/ethereum/ethSchema.js
@@ -61,6 +61,39 @@ export const EtherscanGetTransactions = {
   required: ['result']
 }
 
+export const EtherscanGetInternalTransactions = {
+  type: 'object',
+  properties: {
+    result: {
+      type: 'array',
+      items: {
+        type: 'object',
+        properties: {
+          blockNumber: { type: 'string' },
+          timeStamp: { type: 'string' },
+          hash: { type: 'string' },
+          from: { type: 'string' },
+          to: { type: 'string' },
+          value: { type: 'string' },
+          gas: { type: 'string' },
+          gasUsed: { type: 'string' }
+        },
+        required: [
+          'blockNumber',
+          'timeStamp',
+          'hash',
+          'from',
+          'to',
+          'value',
+          'gas',
+          'gasUsed'
+        ]
+      }
+    }
+  },
+  required: ['result']
+}
+
 export const EtherscanGetTokenTransactions = {
   type: 'object',
   properties: {
@@ -428,6 +461,60 @@ export const AmberdataAccountsTxSchema = {
               'gasPrice',
               'gasUsed',
               'cumulativeGasUsed',
+              'from',
+              'to'
+            ]
+          }
+        }
+      },
+      required: ['records']
+    }
+  },
+  required: ['payload']
+}
+
+export const AmberdataAccountsFuncsSchema = {
+  type: 'object',
+  properties: {
+    payload: {
+      type: 'object',
+      properties: {
+        records: {
+          type: 'array',
+          items: {
+            type: 'object',
+            properties: {
+              transactionHash: { type: 'string' },
+              timestamp: { type: 'number' },
+              blockNumber: { type: 'string' },
+              value: { type: 'string' },
+              initialGas: { type: 'string' },
+              leftOverGas: { type: 'string' },
+              from: {
+                type: 'object',
+                properties: {
+                  address: { type: 'string' }
+                },
+                required: ['address']
+              },
+              to: {
+                type: 'array',
+                items: {
+                  type: 'object',
+                  properties: {
+                    address: { type: 'string' }
+                  },
+                  required: ['address']
+                }
+              }
+            },
+            required: [
+              'transactionHash',
+              'timestamp',
+              'blockNumber',
+              'value',
+              'initialGas',
+              'leftOverGas',
               'from',
               'to'
             ]