+++ src/react-native-io.js
@@ -1,20 +1,5 @@
 // @flow
-
 import { type EdgeCorePluginOptions } from 'edge-core-js/types'
-
-export type FetchJson = (uri: string, opts?: Object) => Object
-
-function makeFetchJson(io): FetchJson {
-  return function fetchJson(uri, opts) {
-    return io.fetch(uri, opts).then(reply => {
-      if (!reply.ok) {
-        throw new Error(`Error ${reply.status} while fetching ${uri}`)
-      }
-      return reply.json()
-    })
-  }
-}
-
 /**
  * Emulates the browser Fetch API more accurately than fetch JSON.
  */
@@ -38,15 +23,9 @@ export function getFetchCors(opts: EdgeCorePluginOptions): Function {
   }
 }
 
-export function getFetchJson(opts: EdgeCorePluginOptions): FetchJson {
-  const nativeIo = opts.nativeIo['edge-currency-accountbased']
-  return nativeIo != null ? nativeIo.fetchJson : makeFetchJson(opts.io)
-}
-
+// TODO: Remove this entire file in the next breaking change.
 export default function makePluginIo() {
   return {
-    fetchJson: makeFetchJson(window),
-
     fetchText(uri: string, opts: Object) {
       return window.fetch(uri, opts).then(reply =>
         reply.text().then(text => ({