+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # edge-currency-accountbased
 
+# 0.6.6 (2020-02-13)
+
+- EOS - Revert fetch update to fix syncing
+
+# 0.6.5 (2020-02-06)
+
+- EOS - Add Greymass Fuel
+
+# 0.6.4 (2020-01-22)
+
+- Add ETH internal transaction support
+
 # 0.6.3 (2020-01-06)
 
 - Add ETHBNT